add card with icons to linkage comp
