+++ app/javascript/app/components/sdg-card/sdg-card-styles.scss
@@ -3,6 +3,7 @@
 .card {
   border: solid 1px $theme-border;
   padding: 15px;
+  position: relative;
 }
 
 .title {
@@ -19,4 +20,14 @@
   border-radius: 5px;
   background-color: $theme-border;
   margin-right: 5px;
+}
+
+.icon {
+  height: 20px;
+  width: 50px;
+  background-repeat: no-repeat;
+  background-size: contain;
+  position: absolute;
+  right: 15px;
+  bottom: 15px;
 }
